requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
requests
validators
aiofiles
dnspython